# backend/src/config/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "E-Commerce MERN API",
    "version": "1.0.0",
    "description": "API documentation for E-Commerce MERN Stack Application (Auto-generated)",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    },
    {
      "url": "https://api.yourdomain.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/wishlist/{productId}": {
      "post": {
        "tags": [
          "Wishlist"
        ],
        "summary": "Add product to wishlist",
        "description": "",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "example": "any"
                  },
                  "comment": {
                    "example": "any"
                  },
                  "rating": {
                    "example": "any"
                  },
                  "orderId": {
                    "example": "any"
                  },
                  "images": {
                    "example": "any"
                  },
                  "video": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Wishlist"
        ],
        "summary": "Remove product from wishlist",
        "description": "",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/wishlist/check/{productId}": {
      "get": {
        "tags": [
          "Wishlist"
        ],
        "summary": "Check if product is in wishlist",
        "description": "",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/wishlist/move-to-cart/{productId}": {
      "post": {
        "tags": [
          "Wishlist"
        ],
        "summary": "Move product from wishlist to cart",
        "description": "",
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your JWT token"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}